requirements_finetune.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
pillow>=10.0.0
numpy>=2.0.0
tqdm>=4.65.0
torchvision>=0.15.0
xformers>=0.0.20
